Issue #182 - Parameterize the location of firebird.msg
firebird.msg is included inside the module dir, and for most of the
cases, it is enough, since the node_modules follow the application.
But in some specific cases, such as on packing the application with
Electron (and probably any Webpack packaging), the file is not
included by default. Moreover, in electron, there are restrictions
on where the file can be placed. An extra option was included to
define where the firebird.msg is placed. By default, the file will be
load from the same old place.